[functional tests] add account configuration
This adds a new config option that allows users to define accounts used in transactions:
```
//! account: <name>, <balance>, [sequence_number]
```
Right now this has limited usage: the transaction will be signed by the default account named "Alice" (case insensitive).
Other accounts are parsed and added to the fake storage. (This is a preparation for running multiple transactions.)