[parser] parse standalone module as program with default main function that does nothing
This allows a standalone module to be parsed:
```
module M {
   ...
}
```
is equivalent to
```
modules:
module M {
   ...
}

script:
main() {
    return;
}
```

We are migrating to a world where each normal transaction is either executing a script, or publishing a module. This syntax change allows us to start splitting existing tests into multiple transactions in parallel to actually implementing the separation.

Once we have split module publishing from running scripts, we'll revisit the parser and forbids the invalid constructions.