[types] Mask (prop)test-only modules behind a conditional compilation feature
## This diff modifies this behaviour:
Test-related terms (including proptest strategies and `Arbitrary` implementations) that are not tests but are still only relevant to tests are put under the `#cfg[any(test, feature = "testing")]` flag. The "testing" feature flag is necessary because `test` does not carry across the crate boundary, whereas a feature does. It is used so that test dependencies import external crates with the `testing` feature.

This continues the trend of D16150452 (where I introduced a smaller version of this bulkhead to constrain access to  powerful `Clone` implementations related to cryptographic key material).

As a byproduct of [cargo issue #291](https://github.com/rust-lang/cargo/issues/2911), and as we need a feature activated through a `Cargo.toml` to access (prop)test members in another crate, I've had to move the *integration* tests at `language/vm/serialization_tests.rs` to their own crate (akin to `testsuite`).

When wise + relevant, crate-wide access to (prop)-test related terms of other crates can be set up, as done in `testsuite/libra_fuzzer`. This should be done (very) sparsely.

## Why this is better:
- less test-related code in our binaries,
- framework to ensure no accidental access to test-related code / helpers outside tests

## Why this is worse:
- slightly more complex `Cargo.toml` files,
- most now need to read the documentation of [`cfg_attr`](https://doc.rust-lang.org/1.0.0/book/conditional-compilation.html#cfg_attr)
- **integration** tests that use conditional members cannot be set up in the `tests/` directory of the crate they test (cargo bug).

## Future improvements:
- extend this masking to more crates