[VM] Adding Script and Move transactions in the VM
This adds execution for Scripts and Modules transactions.
There is minimal testing for it and those code paths should be mostly
closed for testnet.
Next PR is going to move most or all the tests to use the Script and Module
transactions and that will give us confidence that all is good.
We have not removed the Program code yet and left it untouched on purpose.
We will remove it entirely once we remove Program transactions.