[Consensus] [small] Fix vote aggregation
Summary:
We used to aggregate votes based on their vote msg hash, which does not cover the content
of ledger info fields. Hence, we didn't have a guarantee that the LedgerInfos aggregated for
the QC are actually the same. As a result, an honest validator could be tricked to forming a malformed
QC, which would not pass verification.
In this change we're changing the aggregation to be based on the hash of the ledger info.

Testing: added a unit test verifying the aggregation based on the LI hash

Issues: ref #402