[consensus] update safety rules documentation and add TODOs
The documentation had fallen out of date as the code moved from within
consensus, to its own file, to its own crate.
In addition, there were race conditions due to concurrent work happening
in the code base. Listing TODOs to better help track these. So ideally
if a PR were to hit on a TODO it would either be addressed or updated
to reflect the new world order.

Closes: #1713
Approved by: zekun000