[Storage] Add verification when restoring jellyfish-merkle tree
Previously `add_chunk` assumes that input has been validated. Now the
method takes an additional proof and does verification before writing
things to storage.

The basic idea is that we can combine all the known accounts (from the
leftmost one to the most recently added one) and the siblings in the
proof to compute the root hash.

Closes: #1736
Approved by: msmouse