[Consensus] Fix a flaky aggregate_timeout_votes test
Summary:
We recently removed the customized voting power from the SMR tests so they had to be updated.
The specific problem in the `aggregate_timeout_votes` test is that we need to ensure nodes 1 and 2 both receive votes from node 0.
In order to ensure that we need to wait for two votes from 0 upon timeout. Hence, we need to disable communication between 1 and 2 to not interfere with the counting.

Testing: run it in a loop on my laptop, works fine.

Closes: #1986
Approved by: zekun000