[move source language] Use Sequence semicolon source location
In the case where a Sequence is missing the final expression, and where
the Sequence had a final semicolon, attach the source location for the
semicolon to the Exp_::Unit value for the final expression.

This fixes a bunch of error messages where the source location previously
used the span of the entire sequence. There is more churn in the
expected error messages because this causes the messages to be sorted
in a different order.

Closes: #2596
Approved by: tnowacki