[config] Simplify genesis blob output
On startup, we print the current config, this results in a massive dump
of the genesis blob that is largely human unreadable. Until we get a
chance to make a pretty genesis printer, let's hold off on printing
anything other than whether or not it is defined.

running this within a libra-node results in this being displayed:
execution: ExecutionConfig { genesis: Some(...), genesis_file_location: "genesis.blob" },

Closes: #2859
Approved by: mimoo