[key-manager] build rotation transaction builder
This builds the rotation transaction framework and a means to test it
end to end. This also validates recent work on retrieving
ValidatorConfig and ValidatorSet from storage.

The one thing I wonder about writing this is that there are a lot of
places in the code that want executor and/or storage but none of them
have them as conveniently as this. For example, executor tests heavily
depend on the grpc interface, yuck. JSON-RPC built a compatible storage
engine rather than use LibraDB which is really cool but won't work with
executor as it currently stands -- at least not in an easy way. So here
we have yet another approach. I guess once we start cleaning up
executor, we could drag the JSON-RPC storage into that code base and
really simplify testing of these things.

Closes: #2966
Approved by: zekun000