[bitvec] Use Vec<u8> instead of Bytes to avoid dependency
Bytes provides COW (copy-on-write) semantics and faster
serialization/deserialization by consuming the underlying byte buffer in
one shot instead of one-by-one as a sequence. However, since bitvec is
not expected to be used in perf-critical code paths, it's ok for now to
use Vec.

Another alternative for fast serde is `serde_bytes` which can be
explored in the future if the need for more performant
serialization/deserialization arises.

Closes: #2982
Approved by: davidiw