[secure-storage] Add namespace / prefix ability to vault
Per @mgorven's request, Vault now supports different prefixes or
namespaces. The only caveat that the default policy can be and must be
mutated by across these if we want to allow someone to manipulate the
any. Unlikely to be an issue in production.

All secrets are now stored in namespace/key
All policies are now stored in namespace/key
In the case of policies this doesn't actually create a subdirectory, it
is just for convenience so we have the same naming as secrets.

Also worth calling out that the reset macro required to support
recursing into subdirectories.

Closes: #3031
Approved by: mimoo