[key-manager] Implement a loop / reliability mechanism
This introduces a basic event loop for handling key rotation. It detects
stalled transactions and resubmits them. It can handle recovery from
a reboot. It can handle regular rotations. It does not have a config, it
does not have logging, it does not have a main function.

Closes: #3144
Approved by: zekun000