[config] Add clearer and patterned error handling
anyhow is great to throw something together, but without diligence it
becomes pretty useless except for error propagation. Specifically, it is
super easy to just add `?` or `map_err(|e| e.into())` everywhere and not
really annotate errors with enough information. Secondly, even if
properly annotated, the error messages will not be uniform, so what
might be an IO error in one context could be listed as a access error
elsewhere but still fundamentally the same thing. Again, anyhow supports
intelligent usage, but it only comes with discipline and there's no way
to enforce it after leaving a crate. The better solution for strictness
is to use a dedicated error type as done in this PR.

The big win of this PR is that when something bad happens, it should be
very clear where that bad thing is happening... no longer will we get
unclear errors at least from within the config folder.

Also Rust is shortly going to enable backtraces in errors, that will
make it even better... oh happy day...

Closes: #4642
Approved by: mimoo