[move-prover] Fixing wrong reporting of condition failures.
In a recent PR, we had introduced the concept of condition infos. A env::ConditionInfo is a data structure which describes how we handle condition failures, and which messages are reported for them. This information is represented as a map from source position information of the `ensures ...` etc. in the Move source into ConditionInfo. However, some of those source positions have multiple functions, namely invariants, which are used both as pre and
as post conditions. To deal with this the ConditionInfo was extended to represent these both dual purposes.

Existing tests had the bug slipped into the baseline, which this PR corrects as well.

Closes: #4669
Approved by: kkmc