[safety-rules] allow re-initialization of storage
In general, we wanted to avoid re-initialization of storage because it
shouldn't happen in integration and unit tests. However, cluster tests
do expect to re-initialize and we could even imagine cases where some
integration and smoke tests would simply re-initialize as well --
obviously we have worked around that issue to get the testability that
we have, but ... let's just get this working so we can ensure we have a
robust solution on our hands!

Closes: #5144
Approved by: joshlind