[logging] Allow text logs to be replaced with struct logs
This approach opens the `log` field previously only used by text logs
that were converted to struct logs, by struct logs.  In that way, we can
use a category and name rather than a pattern to find log lines
associated in ELK.

The main purpose of this is to have text logs have explicit indexing
guidelines, on how and what we want to index.  The automatic conversion
of text logs was easy, but caused a lot of confusion for anonymous field
names e.g. _0 and _1, and it forced one type of output `debug`.  This
also kills some unused fields, and cleans up the interface to make a
clear difference between mutref and non-mutref functions.

Closes: #5627