[libra-node] Enable running a single node cluster from libra-node
As part of the work to make Libra easier to use, this makes it so we can
start a single node cluster. The next step is to add this to docker.
Then we make a docker-compose that will instantiate a pfn talking to a vfn talking to a
validator and a faucet talking to the vfns json-rpc endpoint  as well.
At which point we can update all the documentation that talks about
running your own testnet to use this instead of the non-education
libra-swarm

Note, this also introduces a really awful file-based logger for the
single node. Could it be better? Hell yeah, is it good eough fo' sho