[move-prover][stdlib] Abstract FixedPoint32, avoid non-linear arithmetics.
This PR introduces an abstraction for the module FixedPoint32 with the goal to avoid non-linear arithmetics in verification conditions. It appears that those are a major obstacle for the SMT solver.

Instead of arbitrary rationals, we use fixed ones for the value of 0.5 and 1.0. The abstraction is performed in a way that the abortion behavior is preserved. Therefore verification of callers is orthogonal (as long as it does not reflect directly over FixedPoint32 values). More precisely, if f is a function in FP32, and A the abstraction, then we have A(f(x)) ~ A(f)(A(x)) (considering aborts as special values).

All existing specifications verify with the abstraction (specifically, no abort conditions are missed, which the prover would complain about). A few timeouts around non-linear arithmetics apparently can be removed. Unfortunately, `LibraAccount::unstaple_lbr` does still timeout.

The specification (see [here](https://github.com/wrwg/libra/blob/abstract-fixedpoint/language/stdlib/modules/doc/FixedPoint32.md)) makes use of the `[abstract]` and `[concrete]` properties of conditions in a non-trivial way for the first time. It appeared necessary to enable the syntax `include [property] Schema` to make this approach practical, so this was added as a side-effect of this PR.

Closes: #6261