[Consensus] Verify vote data earlier
The insert_vote function claims that "Insert a vote and if the vote
is valid, return a QuorumCertificate preferentially over
a TimeoutCertificate if either can can be formed", however the
verification of the validity of the vote data was not performed.

This commit is adding verification of the vote data in the handling
of the vote verification. Notice that this will show up in the logs
as a security event of type "invalid consensus message" through the
epoch_manager message handling.

Closes: #6149