[move-prover] Add support for Boogie :print attribute.
This PR adds support for the new Boogie feature `{:debug ..}`. This is currently only supported in the byte code generator, not yet in boogie_wrapper for model analysis.

The feature is off by default. To turn it on, set `backend.use_boogie_debug_attrib` to true (in cli.rs or via config file or command line).

Closes: #6373